requirements.txt



python-socketio==5.4.0
python-engineio==4.2.1